# api.yaml
openapi: 3.0.3
info:
  title: SittingSpotDataLayer
  description: Data Layer managing internal storage for sitting spots.
  version: 1.0.0
servers:
  - url: https://sitting-spot-dl/api/v1
paths:
  /:
    get:
      summary: Get list of sitting spots
      description: Get list of sitting spots
      operationId: getSittingSpots
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SittingSpotOutDTO'
    post:
      summary: Add a new sitting spot
      description: Add a new sitting spot
      operationId: addSittingSpot
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SittingSpotInDTO'
        required: true
      responses:
        201:
          description: Sitting spot created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SittingSpotOutDTO'
        400:
          description: Invalid input
  /{Id}:
    get:
      summary: Find sitting spot by ID
      description: Returns a single sitting spot
      operationId: getSittingSpotById
      parameters:
        - name: Id
          in: path
          description: ID of the sitting spot
          required: true
          schema:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SittingSpotOutDTO'
        400:
          description: Invalid ID supplied
        404:
          description: Sitting spot not found
    put:
      summary: Updates a sitting spot
      description: Update a sitting spot
      operationId: updateSittingSpot
      requestBody: 
        description: Custom labels to be added to the sitting spot
        content: 
          application/json:
            schema:
              type: array
              items:
                type: string
      responses:
        200:
          description: Sitting spot updated
        404:
          description: Sitting spot not found
  /find:
    get:
      summary: Find list of sitting spots
      description: Givent search parameters return relative sitting spots
      operationId: findSittingSpots
      parameters:
        - name: x
          in: query
          required: true
          schema:
            type: number
        - name: y
          in: query
          required: true
          schema:
            type: number
        - name: area
          in: query
          required: true
          schema:
            type: number
        - name: tags
          in: query
          description: Tags of the sitting spot
          required: false
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Tag'
        - name: labels
          in: query
          description: Labels of the sitting spot
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        200:
          description: Succesful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SittingSpotOutDTO'
components:
  schemas:
    Location:
      description: x,y coordinates
      properties:
        x:
          type: number
        y:
          type: number
    Area:
      description: An area with a location as center
      properties:
        center:
          $ref : '#/components/schemas/Location'
        range:
          type: number
    Tag:
      properties:
        key:
          type: string
        value:
          type: string
    SittingSpotInDTO:
      type: object
      properties:
        id:
          type: string
        location:
          $ref: '#/components/schemas/Location'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
    SittingSpotOutDTO:
      type: object
      properties:
        id:
          type: string
        location:
          $ref: '#/components/schemas/Location'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        labels:
          type: array
          items:
            type: string
    SittingSpot:
      type: object
      properties:
        id:
          type: string
        location:
          $ref : '#/components/schemas/Location'
        tags:
          type: array
          items:
            $ref : '#/components/schemas/Tag'
        labels:
          type: array
          items:
            type: string